+++ common/src/main/java/com/dudko/blazinghot/content/kinetics/blaze_mixer/BlazeMixerRenderer.java
@@ -1,14 +1,14 @@
 package com.dudko.blazinghot.content.kinetics.blaze_mixer;
 
 import com.dudko.blazinghot.registry.BlazingPartialModels;
-import com.jozufozu.flywheel.backend.Backend;
 import com.mojang.blaze3d.vertex.PoseStack;
 import com.mojang.blaze3d.vertex.VertexConsumer;
 import com.simibubi.create.content.kinetics.base.KineticBlockEntityRenderer;
-import com.simibubi.create.foundation.render.CachedBufferer;
-import com.simibubi.create.foundation.render.SuperByteBuffer;
-import com.simibubi.create.foundation.utility.AnimationTickHolder;
 
+import dev.engine_room.flywheel.api.visualization.VisualizationManager;
+import net.createmod.catnip.animation.AnimationTickHolder;
+import net.createmod.catnip.render.CachedBuffers;
+import net.createmod.catnip.render.SuperByteBuffer;
 import net.minecraft.client.renderer.MultiBufferSource;
 import net.minecraft.client.renderer.RenderType;
 import net.minecraft.client.renderer.blockentity.BlockEntityRendererProvider;
@@ -29,7 +29,7 @@ public class BlazeMixerRenderer extends KineticBlockEntityRenderer<BlazeMixerBlo
 	@Override
 	protected void renderSafe(BlazeMixerBlockEntity be, float partialTicks, PoseStack ms, MultiBufferSource buffer, int light, int overlay) {
 
-		if (Backend.canUseInstancing(be.getLevel())) return;
+		if (VisualizationManager.supportsVisualization(be.getLevel())) return;
 
 		BlockState blockState = be.getBlockState();
 
@@ -37,7 +37,7 @@ public class BlazeMixerRenderer extends KineticBlockEntityRenderer<BlazeMixerBlo
 
 		SuperByteBuffer
 				superBuffer =
-				CachedBufferer.partial(BlazingPartialModels.SHAFTLESS_CRIMSON_COGWHEEL, blockState);
+				CachedBuffers.partial(BlazingPartialModels.SHAFTLESS_CRIMSON_COGWHEEL, blockState);
 		standardKineticRotationTransform(superBuffer, be, light).renderInto(ms, vb);
 
 		float renderedHeadOffset = be.getRenderedHeadOffset(partialTicks);
@@ -45,14 +45,13 @@ public class BlazeMixerRenderer extends KineticBlockEntityRenderer<BlazeMixerBlo
 		float time = AnimationTickHolder.getRenderTime(be.getLevel());
 		float angle = ((time * speed * 6 / 10f) % 360) / 180 * (float) Math.PI;
 
-		SuperByteBuffer poleRender = CachedBufferer.partial(BlazingPartialModels.BLAZE_MIXER_POLE, blockState);
+		SuperByteBuffer poleRender = CachedBuffers.partial(BlazingPartialModels.BLAZE_MIXER_POLE, blockState);
 		poleRender.translate(0, -renderedHeadOffset, 0).light(light).renderInto(ms, vb);
 
 		VertexConsumer vbCutout = buffer.getBuffer(RenderType.cutoutMipped());
-
-		SuperByteBuffer headRender = CachedBufferer.partial(BlazingPartialModels.BLAZE_MIXER_HEAD, blockState);
+		SuperByteBuffer headRender = CachedBuffers.partial(BlazingPartialModels.BLAZE_MIXER_HEAD, blockState);
 		headRender
-				.rotateCentered(Direction.UP, angle)
+				.rotateCentered(angle, Direction.UP)
 				.translate(0, -renderedHeadOffset, 0)
 				.light(light)
 				.renderInto(ms, vbCutout);