+++ common/src/main/java/com/dudko/blazinghot/registry/BlazingBlockEntityTypes.java
@@ -2,10 +2,13 @@ package com.dudko.blazinghot.registry;
 
 import com.dudko.blazinghot.BlazingHot;
 import com.dudko.blazinghot.content.block.modern_lamp.ModernLampBlockEntity;
+import com.dudko.blazinghot.content.casting.casting_depot.CastingDepotBlockEntity;
+import com.dudko.blazinghot.content.casting.casting_depot.CastingDepotRenderer;
+import com.dudko.blazinghot.content.kinetics.blaze_mixer.BlazeMixerBlockEntity;
+import com.dudko.blazinghot.content.kinetics.blaze_mixer.BlazeMixerRenderer;
+import com.dudko.blazinghot.content.kinetics.blaze_mixer.BlazeMixerVisual;
 import com.simibubi.create.foundation.data.CreateRegistrate;
 import com.tterrag.registrate.util.entry.BlockEntityEntry;
-
-import dev.architectury.injectables.annotations.ExpectPlatform;
 
 public class BlazingBlockEntityTypes {
 
@@ -18,12 +21,25 @@ public class BlazingBlockEntityTypes {
 					.validBlocks(BlazingBlocks.modernLamps())
 					.register();
 
+	public static final BlockEntityEntry<BlazeMixerBlockEntity>
+			BLAZE_MIXER =
+			REGISTRATE
+					.blockEntity("blaze_mixer", BlazeMixerBlockEntity::of)
+					.visual(() -> BlazeMixerVisual::new)
+					.validBlocks(BlazingBlocks.BLAZE_MIXER)
+					.renderer(() -> BlazeMixerRenderer::new)
+					.register();
+
+	public static final BlockEntityEntry<CastingDepotBlockEntity>
+			CASTING_DEPOT =
+			REGISTRATE
+					.blockEntity("casting_depot", CastingDepotBlockEntity::of)
+					.validBlocks(BlazingBlocks.CASTING_DEPOT)
+					.renderer(() -> CastingDepotRenderer::new)
+					.register();
+
 	public static void register() {
-		platformRegister();
+
 	}
 
-	@ExpectPlatform
-	public static void platformRegister() {
-		throw new AssertionError();
-	}
 }