+++ fabric/src/main/java/com/dudko/blazinghot/config/fabric/BlazingConfigsImpl.java
@@ -4,10 +4,10 @@ import java.util.Map;
 
 import com.dudko.blazinghot.BlazingHot;
 import com.dudko.blazinghot.config.BlazingConfigs;
-import com.simibubi.create.foundation.config.ConfigBase;
 
 import fuzs.forgeconfigapiport.api.config.v2.ForgeConfigRegistry;
 import fuzs.forgeconfigapiport.api.config.v2.ModConfigEvents;
+import net.createmod.catnip.config.ConfigBase;
 import net.minecraftforge.fml.config.ModConfig;
 
 public class BlazingConfigsImpl {