+++ forge/src/main/java/com/dudko/blazinghot/content/entity/forge/BlazeArrowEntityImpl.java
@@ -1,5 +1,7 @@
 package com.dudko.blazinghot.content.entity.forge;
 
+import javax.annotation.ParametersAreNonnullByDefault;
+
 import com.dudko.blazinghot.content.entity.BlazeArrowEntity;
 
 import net.minecraft.world.entity.EntityType;
@@ -8,6 +10,7 @@ import net.minecraft.world.entity.projectile.AbstractArrow;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.level.Level;
 
+@ParametersAreNonnullByDefault
 public class BlazeArrowEntityImpl extends BlazeArrowEntity {
 
 	public BlazeArrowEntityImpl(EntityType<? extends AbstractArrow> entityType, Level level) {