+++ changelog.md
@@ -18,6 +18,23 @@ Create: Blazing Hot 0.7
 * Fixed missing texture for Red Small Modern Lamp Panel
 
 ------------------------------------------------------
+Create: Blazing Hot 0.6.5
+------------------------------------------------------
+
+### Changed:
+
+* [Forge] Updated to Create 6.0.1
+
+------------------------------------------------------
+Create: Blazing Hot 0.6.4
+------------------------------------------------------
+
+### Fixed:
+
+* Fixed a crash when joining a world with a working Mixer
+* [Forge] Fixed incorrect melting time for some metal forms
+
+------------------------------------------------------
 Create: Blazing Hot 0.6.3
 ------------------------------------------------------